feat(ui): dynamic window width
Allows setting a function as the menu width, to enable dynamic
floating window widths.

I frequently work with terminal splits in Tmux, and would run into the
following situation:
1. Open NeoVim in full terminal, do some work, mark files w/ Harpoon
2. Split terminal to run a command on the work I just did, and leave it
   running
3. Go back to NeoVim and toggle Harpoon's window to change files

On #3, the width is set on init, so Harpoon's menu is still set
according to the full terminal size and goes off the NeoVim buffer. By
making adjustments to allow setting a function as the width instead, we
can avoid this issue. The usual method of setting the value explicitly
still works.

Example usage that will set the window width to be 80% of the window
width or 60, whichever is greater:
```lua
require("harpoon").setup {
  menu = {
    width = function()
      return math.max(60, math.floor(vim.api.nvim_win_get_width(0) * 0.8))
    end
  }
}

```